requirements.txt
Django==1.9.5
django-appconf==1.0.1
django-imagekit==3.3
pilkit==1.1.13
Pillow==3.2.0
psycopg2==2.6.1
six==1.10.0
